Added 64 bit fix to struct xthreadInfo.
git-svn-id: svn://0x263a.de/tekuum_src@242 3ae3452d-76a7-4814-9b33-b51e13292e42

+++ sys/sys_public.h
@@ -556,7 +556,9 @@ typedef enum {
 typedef struct {
 	const char *	name;
 	int				threadHandle;
-	unsigned long	threadId;
+	// Techyon RB: 64 bit fix, changed long to int
+	unsigned int	threadId;
+	// Techyon END
 } xthreadInfo;
 
 const int MAX_THREADS				= 10;
